Feat: Number 클래스를 LottoNumber 클래스로 변경
